Anpr-Secure-Traceability-System/requirements.txt
ultralytics
easyocr
fastapi
uvicorn
python-multipart
pyjwt
pandas
opencv-python-headless